Fix several timeout issues ( https://svnweb.cern.ch/trac/cactus/ticket/110#comment:8 ):  - Set mErrorCode to boost::asio::error::timed_out on failure in the timeout callback.  - Add timeout check during connection  - Fix setting of the timeouts     [CACTUS SVN 624]

+++ uhal/uhal/include/uhal/ClientInterface.hpp
@@ -209,7 +209,7 @@ namespace uhal
       	A method to retrieve the timeout period currently being used
       	@return the timeout period currently being used in milliseconds
       */
-      virtual uint32_t getTimeoutPeriod();
+      virtual uint64_t getTimeoutPeriod();
 
     private:
